+++ goms-application/src/main/kotlin/com/goms/v2/domain/auth/data/dto/EmailDto.kt
@@ -1,0 +1,5 @@
+package com.goms.v2.domain.auth.data.dto
+
+data class EmailDto(
+    val email: String
+)
