+++ goms-application/src/main/kotlin/com/goms/v2/domain/auth/exception/AuthenticationNotFoundException.kt
@@ -1,0 +1,6 @@
+package com.goms.v2.domain.auth.exception
+
+import com.goms.v2.common.exception.ErrorCode
+import com.goms.v2.common.exception.GomsException
+
+class AuthenticationNotFoundException: GomsException(ErrorCode.AUTHENTICATION_NOT_FOUND)
