+++ goms-application/src/main/kotlin/com/goms/v2/domain/auth/spi/EmailSendPort.kt
@@ -1,0 +1,7 @@
+package com.goms.v2.domain.auth.spi
+
+interface EmailSendPort {
+
+    fun sendEmail(email: String, authCode: String)
+
+}
