--- goms-application/src/main/kotlin/com/goms/v2/domain/email/data/dto/EmailDto.kt
@@ -1,5 +1,0 @@
-package com.goms.v2.domain.email.data.dto
-
-data class EmailDto(
-    val email: String
-)
