--- goms-application/src/main/kotlin/com/goms/v2/domain/email/exception/AuthCodeNotFoundException.kt
@@ -1,6 +1,0 @@
-package com.goms.v2.domain.email.exception
-
-import com.goms.v2.common.exception.ErrorCode
-import com.goms.v2.common.exception.GomsException
-
-class AuthCodeNotFoundException: GomsException(ErrorCode.AUTH_CODE_NOT_FOUND)
