--- goms-application/src/main/kotlin/com/goms/v2/domain/email/exception/EmailSendFailException.kt
@@ -1,6 +1,0 @@
-package com.goms.v2.domain.email.exception
-
-import com.goms.v2.common.exception.ErrorCode
-import com.goms.v2.common.exception.GomsException
-
-class EmailSendFailException: GomsException(ErrorCode.MAIL_SEND_FAIL)
