--- goms-application/src/main/kotlin/com/goms/v2/domain/email/exception/TooManyAuthCodeRequestException.kt
@@ -1,6 +1,0 @@
-package com.goms.v2.domain.email.exception
-
-import com.goms.v2.common.exception.ErrorCode
-import com.goms.v2.common.exception.GomsException
-
-class TooManyAuthCodeRequestException: GomsException(ErrorCode.TOO_MANY_AUTH_CODE_REQUEST)
