--- goms-application/src/main/kotlin/com/goms/v2/domain/email/spi/EmailSendPort.kt
@@ -1,7 +1,0 @@
-package com.goms.v2.domain.email.spi
-
-interface EmailSendPort {
-
-    fun sendEmail(email: String, authCode: String)
-
-}
