+++ goms-domain/src/main/kotlin/com/goms/v2/repository/auth/AuthCodeRepository.kt
@@ -1,0 +1,10 @@
+package com.goms.v2.repository.auth
+
+import com.goms.v2.domain.auth.AuthCode
+
+interface AuthCodeRepository {
+
+    fun save(authCode: AuthCode)
+    fun findByIdOrNull(email: String): AuthCode?
+
+}
