--- goms-domain/src/main/kotlin/com/goms/v2/repository/email/AuthCodeRepository.kt
@@ -1,10 +1,0 @@
-package com.goms.v2.repository.email
-
-import com.goms.v2.domain.email.AuthCode
-
-interface AuthCodeRepository {
-
-    fun save(authCode: AuthCode)
-    fun findByIdOrNull(email: String): AuthCode?
-
-}
