--- goms-domain/src/main/kotlin/com/goms/v2/repository/email/AuthenticationRepository.kt
@@ -1,11 +1,0 @@
-package com.goms.v2.repository.email
-
-import com.goms.v2.domain.email.Authentication
-
-interface AuthenticationRepository {
-
-    fun save(authentication: Authentication)
-    fun existByEmail(email: String): Boolean
-    fun findByIdOrNull(email: String): Authentication?
-
-}
