+++ goms-infrastructure/src/main/kotlin/com/goms/v2/persistence/auth/repository/AuthenticationRedisRepository.kt
@@ -1,0 +1,8 @@
+package com.goms.v2.persistence.auth.repository
+
+import com.goms.v2.persistence.auth.entity.AuthenticationRedisEntity
+import org.springframework.data.repository.CrudRepository
+
+interface AuthenticationRedisRepository: CrudRepository<AuthenticationRedisEntity, String> {
+    fun findByEmail(email: String): AuthenticationRedisEntity
+}
