--- goms-infrastructure/src/main/kotlin/com/goms/v2/persistence/email/entity/AuthCodeRedisEntity.kt
@@ -1,15 +1,0 @@
-package com.goms.v2.persistence.email.entity
-
-import org.springframework.data.annotation.Id
-import org.springframework.data.redis.core.RedisHash
-import org.springframework.data.redis.core.TimeToLive
-import java.util.concurrent.TimeUnit
-
-@RedisHash("auth_code")
-data class AuthCodeRedisEntity(
-    @Id
-    val email: String,
-    val authCode: String,
-    @TimeToLive(unit = TimeUnit.SECONDS)
-    val expiredAt: Int
-)
