--- goms-infrastructure/src/main/kotlin/com/goms/v2/persistence/email/repository/AuthenticationRedisRepository.kt
@@ -1,8 +1,0 @@
-package com.goms.v2.persistence.email.repository
-
-import com.goms.v2.persistence.email.entity.AuthenticationRedisEntity
-import org.springframework.data.repository.CrudRepository
-
-interface AuthenticationRedisRepository: CrudRepository<AuthenticationRedisEntity, String> {
-    fun findByEmail(email: String): AuthenticationRedisEntity
-}
