+++ goms-presentation/src/main/kotlin/com/goms/v2/domain/auth/dto/request/SendEmailHttpRequest.kt
@@ -1,0 +1,5 @@
+package com.goms.v2.domain.auth.dto.request
+
+data class SendEmailHttpRequest(
+    val email: String
+)
