--- goms-presentation/src/main/kotlin/com/goms/v2/domain/email/dto/request/SendEmailHttpRequest.kt
@@ -1,5 +1,0 @@
-package com.goms.v2.domain.email.dto.request
-
-data class SendEmailHttpRequest(
-    val email: String
-)
